requirements.txt
streamlit
openai
python-dotenv
pypdf
tomli
azure-storage-table